running predictions from linear regression

+++ src/glm_incorporation.py
@@ -1,18 +1,31 @@
 import os
-from typing import List, Set
+import logging
+from typing import List, Set, Dict
 
 import numpy as np
 from pyclam import Manifold, Cluster, criterion, Graph
+from sklearn.metrics import roc_auc_score
 
 from src import datasets as chaoda_datasets
 from src.datasets import METRICS
-from src.meta_ml import auc_scores, METHOD_NAMES
+from src.methods import METHODS, METHOD_NAMES
 from src.utils import TRAIN_PATH
 
 CONSTANTS = os.path.join(os.path.dirname(__file__), '..', 'constants.npy')
 NORMALIZE = True
 SUB_SAMPLE = 100_000
-MAX_DEPTH = 20
+MAX_DEPTH = 50
+
+
+def auc_scores(graph: Graph, labels: List[int]) -> List[float]:
+    scores: List[float] = list()
+    for method in METHODS:
+        anomalies: Dict[int, float] = METHODS[method](graph)
+        y_true, y_score = list(), list()
+        [(y_true.append(labels[k]), y_score.append(v)) for k, v in anomalies.items()]
+        scores.append(roc_auc_score(y_true, y_score))
+
+    return scores
 
 
 def ema(current, previous, smoothing=2.0, period=10):
@@ -68,13 +81,13 @@ def subtree_auc(cluster: Cluster) -> np.ndarray:
 
 def select_graph(manifold: Manifold) -> Manifold:
     graph: Set[Cluster] = set()
-    clusters: List[Cluster] = [manifold.root]
+    clusters: List[Cluster] = [child for child in manifold.root.children]
     while clusters:
         new_clusters: List[Cluster] = list()
         for cluster in clusters:
             if cluster.children:
                 # noinspection PyUnresolvedReferences
-                if cluster.chaoda_auc >= np.percentile(subtree_auc(cluster), q=60):
+                if cluster.chaoda_auc >= np.percentile(subtree_auc(cluster), q=75):
                     graph.add(cluster)
                 else:
                     new_clusters.extend(cluster.children)
@@ -83,7 +96,18 @@ def select_graph(manifold: Manifold) -> Manifold:
         clusters = new_clusters
 
     manifold.graph = Graph(*graph)
+
+    # max_depth = max((cluster.depth for cluster in graph))
+    # manifold.layers[max_depth].build_edges()
+
     manifold.graph.build_edges()
+
+    depths = [cluster.depth for cluster in manifold.graph.clusters]
+    subsumed_ratio = len(manifold.graph.subsumed_clusters) / manifold.graph.cardinality
+    logging.info(f'selected graph from auc_predictions:\n'
+                 f'{manifold.graph.cardinality} clusters,\n'
+                 f'depths ({min(depths)}, {max(depths)}),\n'
+                 f'with {subsumed_ratio:.3f} subsumed ratio.')
 
     return manifold
 
@@ -114,7 +138,7 @@ def evaluate_auc(datasets: List[str], metrics: List[str], constants: np.array, f
                 min_points = 16
                 # continue
 
-            manifold = Manifold(data, METRICS[metric]).build_tree(
+            manifold = Manifold(data, METRICS[metric]).build(
                 criterion.MaxDepth(MAX_DEPTH),
                 criterion.MinPoints(min_points),
             )
@@ -131,14 +155,17 @@ def evaluate_auc(datasets: List[str], metrics: List[str], constants: np.array, f
 
 
 if __name__ == "__main__":
+    np.random.seed(42)
+
     os.makedirs(TRAIN_PATH, exist_ok=True)
 
-    _datasets = ['musk']
+    _datasets = ['cardio']
     _metrics = ['euclidean', 'manhattan']
-    _constants = np.array([1, 2, 3, 4, 5, 6])  # np.fromfile(CONSTANTS, dtype=np.float)
+    _constants = np.array([0.000, 0.414, 0.047, -0.000, -0.483, -0.040])  # np.fromfile(CONSTANTS, dtype=np.float)
     _filename = os.path.join(TRAIN_PATH, 'lr_predictions.csv')
-    with open(_filename, 'w') as _fp:
-        _methods = ','.join([f'{_method}' for _method in METHOD_NAMES.values()])
-        _fp.write(f'dataset,metric,{_methods}\n')
+    if not os.path.exists(_filename):
+        with open(_filename, 'w') as _fp:
+            _methods = ','.join([f'{_method}' for _method in METHOD_NAMES.values()])
+            _fp.write(f'dataset,metric,{_methods}\n')
 
     evaluate_auc(_datasets, _metrics, _constants, _filename)
